lots of docs changes
