+++ docs/data_interface/data_interface.md
@@ -1,0 +1,4 @@
+@typedef {Object} can-connect/DataInterface DataInterface
+@parent can-connect.types
+
+@signature `DataInterface`
