+++ docs/data_interface/getListData.md
@@ -1,5 +1,5 @@
-@typedef {function} connection.getListData getListData
-@parent can-connect.data_interface
+@typedef {function} can-connect/connection.getListData connection.getListData
+@parent can-connect/DataInterface
 
 @description Retrieves list data for a particular set.
 
@@ -7,7 +7,7 @@
 
   Returns a promise that resolves to the list data for a particular set.
 
-  @param {Object} set A object that represents the set of data needed to be loaded.
+  @param {can-set/Set} set A object that represents the set of data needed to be loaded.
 
   @return {Promise<can-connect.listData>} A promise that resolves to the [can-connect.listData] format.
 
@@ -15,7 +15,7 @@
 
 ## Use
 
-Extensions like [can-connect/data-url] make it easy to implement `getListData`, but it can be as simple as:
+Extensions like [can-connect/data/url/url] make it easy to implement `getListData`, but it can be as simple as:
 
 ```
 var connection = connect([],{
