+++ docs/data_interface/updateListData.md
@@ -1,5 +1,5 @@
-@typedef {function} connection.updateListData updateListData
-@parent can-connect.data_interface
+@typedef {function} connection.updateListData connection.updateListData
+@parent can-connect/DataInterface
 
 @description Updates list data for a particular set.
 
@@ -8,7 +8,7 @@
   Returns a promise that resolves to the list data for a particular set.
 
   @param {can-connect.listData} listData A object that represents the set of data needed to be loaded.
-  
+
   @param {Object} [set] The set of data that is updating.
 
   @return {Promise<can-connect.listData>} A promise that resolves to the updated [can-connect.listData].
