+++ docs/instance_interface/instance_interface.md
@@ -1,0 +1,6 @@
+@typedef can-connect/InstanceInterface InstanceInterface
+@parent can-connect.types
+
+@signature `InstanceInterface`
+
+Blah
