+++ docs/instance_interface/save.md
@@ -1,2 +1,2 @@
-@function connection.save save
-@parent can-connect.externalCRUD
+@function can-connect/connection.save connection.save
+@parent can-connect/InstanceInterface
