+++ docs/types/instance.md
@@ -1,0 +1,3 @@
+@typedef {Object} can-connect/Instance Instance
+@parent can-connect.types
+@description An instance of some JavaScript type.
