+++ docs/types/list.md
@@ -1,0 +1,3 @@
+@typedef {Object} can-connect.List List
+@parent can-connect.types
+@description A list type.
