+++ docs/types/set.md
@@ -1,0 +1,3 @@
+@typedef {Object} can-set/Set Set
+@parent can-connect.types
+@description The parameters used to retrieve a list.
