+++ src/can/map/define-map-test.js
@@ -5,18 +5,18 @@ var List = require("can-define/list/list");
 var compute = require("can-compute");
 
 // load connections
-require("can-connect/constructor/");
-require("can-connect/can/map/");
-require("can-connect/constructor/store/");
-require("can-connect/data/callbacks/");
-require("can-connect/data/callbacks-cache/");
-require("can-connect/data/combine-requests/");
-require("can-connect/data/localstorage-cache/");
-require("can-connect/data/parse/");
-require("can-connect/data/url/");
-require("can-connect/fall-through-cache/");
-require("can-connect/real-time/");
-require("can-connect/data/inline-cache/");
+var constructor = require("can-connect/constructor/");
+var canMap = require("can-connect/can/map/");
+var constructorStore = require("can-connect/constructor/store/");
+var dataCallbacks = require("can-connect/data/callbacks/");
+var callbacksCache = require("can-connect/data/callbacks-cache/");
+var combineRequests = require("can-connect/data/combine-requests/");
+var localCache = require("can-connect/data/localstorage-cache/");
+var dataParse = require("can-connect/data/parse/");
+var dataUrl = require("can-connect/data/url/");
+var fallThroughCache = require("can-connect/fall-through-cache/");
+var realTime = require("can-connect/real-time/");
+var inlineCache = require("can-connect/data/inline-cache/");
 require("when/es6-shim/Promise");
 
 
@@ -51,7 +51,7 @@ var cleanUndefineds = function(obj) {
 	}
 };
 
-QUnit.module("can-connect/can/map with define",{
+QUnit.module("can-connect/can/map/map with define",{
 	setup: function(){
 
 		var Todo = Map.extend({
@@ -67,7 +67,7 @@ QUnit.module("can-connect/can/map with define",{
 		this.Todo = Todo;
 		this.TodoList = TodoList;
 
-		var cacheConnection = connect(["data-localstorage-cache"],{
+		var cacheConnection = connect([localCache],{
 			name: "todos"
 		});
 		cacheConnection.clear();
@@ -76,17 +76,17 @@ QUnit.module("can-connect/can/map with define",{
 		this.Todo = Todo;
 
 		this.todoConnection = connect([
-			"constructor",
-			"can-map",
-			"constructor-store",
-			"data-callbacks",
-			"data-callbacks-cache",
-			"data-combine-requests",
-			"data-inline-cache",
-			"data-parse",
-			"data-url",
-			"fall-through-cache",
-			"real-time"],
+			constructor,
+			canMap,
+			constructorStore,
+			dataCallbacks,
+			callbacksCache,
+			combineRequests,
+			inlineCache,
+			dataParse,
+			dataUrl,
+			fallThroughCache,
+			realTime],
 			{
 				url: "/services/todos",
 				cacheConnection: cacheConnection,
@@ -173,6 +173,7 @@ QUnit.test("real-time super model", function(){
 		bindFunc = function(){
 			console.log("length changing");
 		};
+
 	Promise.all([connection.getList({type: "important"}), connection.getList({due: "today"})])
 		.then(function(result){
 
@@ -483,17 +484,17 @@ QUnit.test("reads id from set algebra (#82)", function(){
 
 
 	var todoConnection = connect([
-		"constructor",
-		"can-map",
-		"constructor-store",
-		"data-callbacks",
-		"data-callbacks-cache",
-		"data-combine-requests",
-		"data-inline-cache",
-		"data-parse",
-		"data-url",
-		"fall-through-cache",
-		"real-time"],
+		constructor,
+		canMap,
+		constructorStore,
+		dataCallbacks,
+		callbacksCache,
+		combineRequests,
+		inlineCache,
+		dataParse,
+		dataUrl,
+		fallThroughCache,
+		realTime],
 		{
 			url: "/services/todos",
 			Map: Todo,
