+++ src/can/map/map.md
@@ -1,10 +1,10 @@
-@module {connect.Behavior} can-connect/can/map can/map
-@group can-connect/can/map.map-static 0 Map Static Methods
-@group can-connect/can/map.map 1 Map Instance Methods
-@group can-connect/can/map.hydrators 2 Hydrators
-@group can-connect/can/map.serializers 3 Serializers
-@group can-connect/can/map.identifiers 4 Identifiers
-@group can-connect/can/map.instance-callbacks 5 Instance Callbacks
+@module {connect.Behavior} can-connect/can/map/map
+@group can-connect/can/map/map.map-static 0 Map Static Methods
+@group can-connect/can/map/map.map 1 Map Instance Methods
+@group can-connect/can/map/map.hydrators 2 Hydrators
+@group can-connect/can/map/map.serializers 3 Serializers
+@group can-connect/can/map/map.identifiers 4 Identifiers
+@group can-connect/can/map/map.instance-callbacks 5 Instance Callbacks
 @parent can-connect.behaviors
 
 Connects a [can.Map](http://canjs.com/docs/can.Map.html) to everything that needs to be connected to
@@ -16,7 +16,7 @@ Connects a [can.Map](http://canjs.com/docs/can.Map.html) to everything that need
   [can.List](http://canjs.com/docs/can.List.html).
   Adds static and prototype methods to the Map that make use of the connection's
   methods.
-  
+
 @body
 
 ## Use
@@ -48,8 +48,14 @@ var TodoList = can.List.extend({
 Next, pass the Map and List constructor functions to `connect` as options. The following
 creates a connection that connects `Todo` and `TodoList` to a restful URL:
 
-```
-var todoConnection = connect(["data-url","constructor","can/map"],{
+```js
+var connect = require("can-connect");
+
+var todoConnection = connect([
+    require("can-connect/data/url/url"),
+    require("can-connect/constructor/constructor"),
+    require("can-connect/can/map/map")
+],{
   Map: Todo,
   List: TodoList,
   url: "/services/todos"
@@ -65,7 +71,7 @@ todoConnection.getList({}).then(function(todos){
 })
 ```
 
-However, because `can/map` adds methods to the `Map` option, you can use `Todo` directly to 
+However, because `can/map` adds methods to the `Map` option, you can use `Todo` directly to
 CRUD `Todo` and `TodoList`s:
 
 ```
@@ -82,4 +88,3 @@ new Todo({name: "dishes"}).save().then(function(todo){
     });
 });
 ```
-
