+++ src/can/super-map/super-map.js
@@ -1,44 +1,44 @@
 var connect = require("can-connect");
 
-require("../../constructor/");
-require("../map/");
-require("../../constructor/store/");
-require("../../constructor/callbacks-once/");
-require("../../data/callbacks/");
-require("../../data/callbacks-cache/");
-require("../../data/combine-requests/");
-require("../../data/inline-cache/");
-require("../../data/localstorage-cache/");
-require("../../data/parse/");
-require("../../data/url/");
-require("../../fall-through-cache/");
-require("../../real-time/");
+var constructor = require("../../constructor/");
+var canMap = require("../map/");
+var constructorStore = require("../../constructor/store/");
+var dataCallbacks = require("../../data/callbacks/");
+var callbacksCache = require("../../data/callbacks-cache/");
+var combineRequests = require("../../data/combine-requests/");
+var localCache = require("../../data/localstorage-cache/");
+var dataParse = require("../../data/parse/");
+var dataUrl = require("../../data/url/");
+var fallThroughCache = require("../../fall-through-cache/");
+var realTime = require("../../real-time/");
+var inlineCache = require("../../data/inline-cache/");
+var callbacksOnce = require("../../constructor/callbacks-once/");
 
 var $ = require("jquery");
 
 connect.superMap = function(options){
 
 	var behaviors = [
-		"constructor",
-		"can-map",
-		"constructor-store",
-		"data-callbacks",
-		"data-combine-requests",
-		"data-inline-cache",
-		"data-parse",
-		"data-url",
-		"real-time",
-		"constructor-callbacks-once"];
+		constructor,
+		canMap,
+		constructorStore,
+		dataCallbacks,
+		combineRequests,
+		inlineCache,
+		dataParse,
+		dataUrl,
+		realTime,
+		callbacksOnce];
 
 	if(typeof localStorage !== "undefined") {
 		if(!options.cacheConnection) {
-			options.cacheConnection = connect(["data-localstorage-cache"],{
+			options.cacheConnection = connect([localCache],{
 				name: options.name+"Cache",
 				idProp: options.idProp,
 				algebra: options.algebra
 			});
 		}
-		behaviors.push("data-callbacks-cache","fall-through-cache");
+		behaviors.push(callbacksCache,fallThroughCache);
 	}
 	options.ajax = $.ajax;
