+++ src/can/super-map/super-map.md
@@ -1,32 +1,32 @@
-@module {function} can-connect/can/super-map can/super-map
+@module {function} can-connect/can/super-map/super-map
 @parent can-connect.modules
 
-Create connection with many of the best behaviors in can-connect and hook it up to 
+Create connection with many of the best behaviors in can-connect and hook it up to
 a [can.Map](http://canjs.com/docs/can.Map.html).
 
 @signature `superMap(options)`
 
-  Creates a connection with the following behaviors: [can-connect/constructor],
-  [can-connect/can/map],
-  [can-connect/constructor/store],
-  [can-connect/data/callbacks],
-  [can-connect/data/callbacks-cache],
-  [can-connect/data/combine-requests],
-  [can-connect/data/inline-cache],
-  [can-connect/data-parse],
-  [can-connect/data-url],
-  [can-connect/real-time],
-  [can-connect/fall-through-cache],
-  [can-connect/constructor/callbacks-once].
+  Creates a connection with the following behaviors: [can-connect/constructor/constructor],
+  [can-connect/can/map/map],
+  [can-connect/constructor/store/store],
+  [can-connect/data/callbacks/callbacks],
+  [can-connect/data/callbacks-cache/callbacks-cache],
+  [can-connect/data/combine-requests/combine-requests],
+  [can-connect/data/inline-cache/inline-cache],
+  [can-connect/data/parse/parse],
+  [can-connect/data/url/url],
+  [can-connect/real-time/real-time],
+  [can-connect/fall-through-cache/fall-through-cache],
+  [can-connect/constructor/callbacks-once/callbacks-once].
 
-  And creates a [can-connect/data/localstorage-cache] to use as a [connect.base.cacheConnection].
+  And creates a [can-connect/data/localstorage-cache/localstorage-cache] to use as a [can-connect/base/base.cacheConnection].
 
-@body 
+@body
 
 ## Use
 
 The `can-connect/can/super-map` module exports a helper function that creates a connection
-with the "standard" behaviors in can-connect and hooks it up to a 
+with the "standard" behaviors in can-connect and hooks it up to a
 [can.Map](http://canjs.com/docs/can.Map.html) and [can.List](http://canjs.com/docs/can.List.html).
 
 If you are using CanJS, this is an easy way to create a connection that can be useful and
@@ -39,7 +39,7 @@ var Todo = can.Map.extend({ ... });
 var TodoList = can.List.extend({Map: Todo},{ ... });
 ```
 
-Next, call `superMap` with all of the options needed by the behaviors that `superMap` adds: 
+Next, call `superMap` with all of the options needed by the behaviors that `superMap` adds:
 
 ```
 var todoConnection = superMap({
@@ -51,7 +51,7 @@ var todoConnection = superMap({
 });
 ```
 
-As, [can-connect/can/map] adds CRUD methods to the `Map` option, you can use those to create, 
+As, [can-connect/can/map/map] adds CRUD methods to the `Map` option, you can use those to create,
 read, update and destroy todos:
 
 ```
@@ -68,5 +68,3 @@ new Todo({name: "dishes"}).save().then(function(todo){
     });
 });
 ```
-
-
