+++ src/data/callbacks-cache/callbacks-cache.js
@@ -1,10 +1,15 @@
 /**
- * @module can-connect/data/callbacks-cache data-callbacks-cache
+ * @module can-connect/data/callbacks-cache/callbacks-cache
  * @parent can-connect.behaviors
  *
- * Calls [connect.base.cacheConnection] methods whenever
- * the [can-connect/data/callbacks data interface callbacks] are called. This is
- * useful for making sure a [connect.base.cacheConnection] is updated whenever data is updated.
+ * Callback [can-connect/base/base.cacheConnection] methods when [can-connect/data/callbacks/callbacks data interface callbacks]
+ * are called.
+ *
+ * @signature `dataCallbacksCache( baseConnection )`
+ *
+ *   Implements the [can-connect/data/callbacks/callbacks data callbacks] so that a corresponding method is called
+ *   on the [can-connect/base/base.cacheConnection].This is
+ *   useful for making sure a [can-connect/base/base.cacheConnection] is updated whenever data is updated.
  */
 var connect = require("can-connect");
 var assign = require("can-util/js/assign/assign");
@@ -13,27 +18,27 @@ var each = require("can-util/js/each/each");
 // wires up the following methods
 var pairs = {
 	/**
-	 * @function can-connect/data/callbacks-cache.createdData createdData
-	 * @parent can-connect/data/callbacks-cache
+	 * @function can-connect/data/callbacks-cache/callbacks-cache.createdData createdData
+	 * @parent can-connect/data/callbacks-cache/callbacks-cache
 	 *
 	 * Called with the resolved response data
-	 * of [connection.createData]. Calls `createData` on the [connect.base.cacheConnection].
+	 * of [can-connect/connection.createData]. Calls `createData` on the [can-connect/base/base.cacheConnection].
 	 */
 	createdData: "createData",
 	/**
-	 * @function can-connect/data/callbacks-cache.updatedData updatedData
-	 * @parent can-connect/data/callbacks-cache
+	 * @function can-connect/data/callbacks-cache/callbacks-cache.updatedData updatedData
+	 * @parent can-connect/data/callbacks-cache/callbacks-cache
 	 *
 	 * Called with the resolved response data
-	 * of [connection.updateData]. Calls `updateData` on the [connect.base.cacheConnection].
+	 * of [can-connect/connection.updateData]. Calls `updateData` on the [can-connect/base/base.cacheConnection].
 	 */
 	updatedData: "updateData",
 	/**
-	 * @function can-connect/data/callbacks-cache.destroyedData destroyedData
-	 * @parent can-connect/data/callbacks-cache
+	 * @function can-connect/data/callbacks-cache/callbacks-cache.destroyedData destroyedData
+	 * @parent can-connect/data/callbacks-cache/callbacks-cache
 	 *
 	 * Called with the resolved response data
-	 * of [connection.destroyData]. Calls `destroyData` on the [connect.base.cacheConnection].
+	 * of [can-connect/connection.destroyData]. Calls `destroyData` on the [can-connect/base/base.cacheConnection].
 	 */
 	destroyedData: "destroyData"
 	//gotInstanceData: "updateListData"
@@ -41,7 +46,7 @@ var pairs = {
 
 
 
-module.exports = connect.behavior("data-callbacks-cache",function(baseConnect){
+module.exports = connect.behavior("data/callbacks-cache",function(baseConnect){
 
 	var behavior = {};
