+++ src/data/worker/demo/todo_connection.js
@@ -1,10 +1,10 @@
 import connect from "can-connect";
-import "can-connect/data/url/";
-import "can-connect/constructor/";
-import "can-connect/constructor/store/";
+import dataUrl from "can-connect/data/url/";
+import constructor from "can-connect/constructor/";
+import constructorStore from "can-connect/constructor/store/";
 import "can-connect/data/memory-cache/";
-import "can-connect/cache-requests/";
-import "can-connect/data/worker/";
+import cacheRequests from "can-connect/cache-requests/cache-requests/";
+import dataWorker from "can-connect/data/worker/";
 import fixture from "can-fixture";
 
 var cache = connect(['data-memory-cache'],{
@@ -17,11 +17,11 @@ if(typeof document !== "undefined") {
 }
 
 var todosConnection = connect([
-	"data-url",
-	"cache-requests",
-	"data-worker",
-	"constructor",
-	"constructor-store"],
+	dataUrl,
+	cacheRequests,
+	dataWorker,
+	constructor,
+	constructorStore],
   {
     url: "/todos",
     cacheConnection: cache,
