+++ src/data/worker/worker.md
@@ -1,17 +1,17 @@
-@module can-connect/data/worker data-worker
+@module can-connect/data/worker/worker
 @parent can-connect.behaviors
-@group can-connect/data/worker.identifiers Identifiers
-@group can-connect/data/worker.data Data Methods
+@group can-connect/data/worker/worker.identifiers Identifiers
+@group can-connect/data/worker/worker.data Data Methods
 
 @signature `dataWorker(baseConnection)`
 
 If a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Using_web_workers)
-is provided, overwrites the "data interface methods" to package the arguments and send them as 
+is provided, overwrites the "data interface methods" to package the arguments and send them as
 part of a `postMessage` to the Worker.
 
 
 If a `Worker` is not provided, it is assumed "data-worker" is being added
-within a worker thread.  It listens to messages sent to the Worker, calls the specified "data interface method" 
+within a worker thread.  It listens to messages sent to the Worker, calls the specified "data interface method"
 and sends a message back with the result.
 
 Any data methods called on the `window` connection will wait until the `worker` connection
@@ -21,16 +21,16 @@ has established a handshake.
 
 ## Use
 
-The best way to use `data-worker` is to create a connection module that works when loaded in 
-either the `window` or in a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Using_web_workers). 
+The best way to use `data-worker` is to create a connection module that works when loaded in
+either the `window` or in a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Using_web_workers).
 This pattern tends to work even if workers are not supported.
 
-The following creates a connection that does the work of [can-connect/cache-requests], 
-[can-connect/data-url], and [can-connect/data/memory-cache] in a worker thread.  
+The following creates a connection that does the work of [can-connect/cache-requests/cache-requests],
+[can-connect/data/url/url], and [can-connect/data/memory-cache/memory-cache] in a worker thread.  
 
 @demo can-connect/src/data/worker/demo/worker.html
 
-The `todo_connection` module can be found [here](https://github.com/canjs/can-connect/blob/master/src/data/worker/demo/todo_connection.js) 
+The `todo_connection` module can be found [here](https://github.com/canjs/can-connect/blob/master/src/data/worker/demo/todo_connection.js)
 and looks like the following:
 
 
@@ -68,7 +68,7 @@ var todoConnection = connect([
 The things to notice:
 
 1. A `Worker` should be passed as the `worker` option
-that loads a connection with the same name as the connection in the `window`.  In thise case, the same 
+that loads a connection with the same name as the connection in the `window`.  In thise case, the same
 connection module is loaded so everything works.
 
 2. A single `Worker` could load multiple connection modules and perform other behavior.  
@@ -110,11 +110,5 @@ var todoConnection = connect([
   });
 ```
 
-However, the problem with the two-module approach is that it will not work 
+However, the problem with the two-module approach is that it will not work
 if Workers are not supported by your browser.
-
-
-
-
-
-
