+++ src/helpers/weak-reference-map.js
@@ -1,5 +1,35 @@
 var assign = require("can-util/js/assign/assign");
 
+/**
+ * @module {function} can-connect/helpers/weak-reference-map WeakReferenceMap
+ * @parent can-connect.modules
+ *
+ * Provides a map that only contains keys that are referenced.
+ *
+ * @signature `new WeakReferenceMap()`
+ *
+ *   Creates a new weak reference map.
+ *
+ * @body
+ *
+ * ## Use
+ *
+ * ```
+ * var WeakReferenceMap = require("can-connect/helpers/weak-reference-map");
+ * var wrm = new WeakReferenceMap();
+ * var task1 = {id: 1, name: "do dishes"};
+ *
+ * wrm.addReference("1", task1);
+ * wrm.has("1") //-> true
+ * wrm.addReference("1", task1);
+ * wrm.has("1") //-> true
+ * wrm.removeReference("1", task1);
+ * wrm.has("1") //-> true
+ * wrm.removeReference("1", task1);
+ * wrm.has("1") //-> false
+ * ```
+ */
+
 var WeakReferenceMap = function(){
 	this.set = {};
 };
@@ -7,10 +37,32 @@ var WeakReferenceMap = function(){
 // if weakmap, we can add and never worry ...
 // otherwise, we need to have a count ...
 
-assign(WeakReferenceMap.prototype,{
+assign(WeakReferenceMap.prototype,
+/**
+ * @prototype
+ */
+	{
+	/**
+	 * @function can-connect/helpers/weak-reference-map.prototype.has has
+	 * @signature `weakReferenceMap.has(key)`
+	 *
+	 *   Returns if key is in the set.
+	 *
+	 *   @param  {String} key A key to look for.
+	 *   @return {Boolean} If the key exists.
+	 */
 	has: function(key){
 		return !!this.set[key];
 	},
+	/**
+	 * @function can-connect/helpers/weak-reference-map.prototype.addReference addReference
+	 * @signature `WeakReferenceMap.addReference(key, item)`
+	 *
+	 *   Adds a reference to item as key and increments the reference count. This should be called
+	 *   when a value should be managed by something, typically the [can-connect/constructor/store/store].
+	 *
+	 *   @param  {String} key The key of the item in the store.
+	 */
 	addReference: function(key, item){
 		var data = this.set[key];
 		if(!data) {
@@ -22,6 +74,15 @@ assign(WeakReferenceMap.prototype,{
 		}
 		data.referenceCount++;
 	},
+	/**
+	 * @function can-connect/helpers/weak-reference-map.prototype.deleteReference deleteReference
+	 * @signature `weakReferenceMap.deleteReference(key)`
+	 *
+	 *   Decrements the reference count for key and removes it if the reference count is `0`. This should be called
+	 *   when a value should not be managed by something, typically the [can-connect/constructor/store/store].
+	 *
+	 *   @param  {String} key The key of the item in the store.
+	 */
 	deleteReference: function(key){
 		var data = this.set[key];
 		if(data){
@@ -31,12 +92,29 @@ assign(WeakReferenceMap.prototype,{
 			}
 		}
 	},
+	/**
+	 * @function can-connect/helpers/weak-reference-map.prototype.get get
+	 * @signature `weakReferenceMap.get(key)`
+	 *
+	 *   Returns the value stored at key if it's in the store.
+	 *
+	 *   @param  {String} key The key of the item in the store.
+	 *   @return {*|undefined} The item if it's available.
+	 */
 	get: function(key){
 		var data = this.set[key];
 		if(data) {
 			return data.item;
 		}
 	},
+	/**
+	 * @function can-connect/helpers/weak-reference-map.prototype.forEach forEach
+	 * @signature `weakReferenceMap.forEach(callback)`
+	 *
+	 *   Calls `callback` for every value in the store.
+	 *
+	 *   @param  {function(*,String)} callback(item,key) A callback handler.
+	 */
 	forEach: function(cb){
 		for(var id in this.set) {
 			cb(this.set[id].item, id);
