+++ src/real-time/real-time.js
@@ -1,8 +1,8 @@
 /**
- * @module can-connect/real-time real-time
+ * @module can-connect/real-time/real-time
  * @parent can-connect.behaviors
- * @group can-connect/real-time.methods 0 Methods
- * @group can-connect/real-time.callbacks 1 Data Callbacks
+ * @group can-connect/real-time/real-time.methods 0 Methods
+ * @group can-connect/real-time/real-time.callbacks 1 Data Callbacks
  *
  * Update lists to include or exclude instances based
  * on set logic.
@@ -10,16 +10,16 @@
  * @signature `realTime( baseConnection )`
  *
  *   Overwrites the "data callback" methods and provides
- *   [can-connect/real-time.createInstance],
- *   [can-connect/real-time.updateInstance], and
- *   [can-connect/real-time.destroyInstance] methods
+ *   [can-connect/real-time/real-time.createInstance],
+ *   [can-connect/real-time/real-time.updateInstance], and
+ *   [can-connect/real-time/real-time.destroyInstance] methods
  *   that
  *   update lists to include or exclude a created,
  *   updated, or destroyed instance.
  *
  *   An instance is put in a list if it is a
  *   [set.subset](https://github.com/canjs/can-set#setsubset)
- *   of the [connect.base.listSet].
+ *   of the [can-connect/base/base.listSet].
  *
  *   Currently, all items are added at the end of the list
  *   until [set.add](https://github.com/canjs/can-set/issues/2)
@@ -52,9 +52,9 @@
  * })
  * ```
  *
- * Finally, use one of the  [can-connect/real-time.createInstance],
- * [can-connect/real-time.updateInstance], and
- * [can-connect/real-time.destroyInstance] methods to tell the connection
+ * Finally, use one of the  [can-connect/real-time/real-time.createInstance],
+ * [can-connect/real-time/real-time.updateInstance], and
+ * [can-connect/real-time/real-time.destroyInstance] methods to tell the connection
  * that data has changed.  The connection will update (by calling splice)
  * each list accordingly.
  *
@@ -137,21 +137,21 @@ var indexOf = require("can-connect/helpers/get-index-by-id");
 module.exports = connect.behavior("real-time",function(baseConnect){
 	return {
 		/**
-		 * @function can-connect/real-time.createInstance createInstance
-		 * @parent can-connect/real-time.methods
+		 * @function can-connect/real-time/real-time.createInstance createInstance
+		 * @parent can-connect/real-time/real-time.methods
 		 *
 		 * Programatically indicate a new instance has been created.
 		 *
 		 * @signature `connection.createInstance(props)`
 		 *
 		 *   If there is no instance in the [can.connect/constructor-store.instanceStore]
-		 *   for `props`'s [connect.base.id], an instance is [can-connect/constructor.hydrateInstance hydrated],
-		 *   added to the store, and then [can-connect/real-time.createdData] is called with
-		 *   `props` and the hydrated instance's serialized data. [can-connect/real-time.createdData]
+		 *   for `props`'s [can-connect/base/base.id], an instance is [can-connect/constructor.hydrateInstance hydrated],
+		 *   added to the store, and then [can-connect/real-time/real-time.createdData] is called with
+		 *   `props` and the hydrated instance's serialized data. [can-connect/real-time/real-time.createdData]
 		 *   will add this instance to any lists the instance belongs to.
 		 *
 		 *   If this instance has already been created, calls
-		 *   [can-connect/real-time.updateInstance] with `props`.
+		 *   [can-connect/real-time/real-time.updateInstance] with `props`.
 		 *
 		 *   @param {Object} props
 		 *
@@ -196,8 +196,8 @@ module.exports = connect.behavior("real-time",function(baseConnect){
 			}
 		},
 		/**
-		 * @function can-connect/real-time.createdData createdData
-		 * @parent can-connect/real-time.callbacks
+		 * @function can-connect/real-time/real-time.createdData createdData
+		 * @parent can-connect/real-time/real-time.callbacks
 		 *
 		 * Called whenever instance data is created.
 		 *
@@ -211,7 +211,7 @@ module.exports = connect.behavior("real-time",function(baseConnect){
 		 *   Next, it goes through every list in the [can.connect/constructor-store.listStore],
 		 *   test if the instance's data belongs in that list.  If it does,
 		 *   adds the instance's data to the serialized list data and
-		 *   [can-connect/constructor.updatedList updates the list].
+		 *   [can-connect/constructor/constructor.updatedList updates the list].
 		 *
 		 *
 		 *
@@ -231,8 +231,8 @@ module.exports = connect.behavior("real-time",function(baseConnect){
 			return undefined;
 		},
 		/**
-		 * @function can-connect/real-time.updatedData updatedData
-		 * @parent can-connect/real-time.callbacks
+		 * @function can-connect/real-time/real-time.updatedData updatedData
+		 * @parent can-connect/real-time/real-time.callbacks
 		 *
 		 * Called whenever instance data is updated.
 		 *
@@ -256,14 +256,14 @@ module.exports = connect.behavior("real-time",function(baseConnect){
 			return undefined;
 		},
 		/**
-		 * @function can-connect/real-time.updateInstance updateInstance
-		 * @parent can-connect/real-time.methods
+		 * @function can-connect/real-time/real-time.updateInstance updateInstance
+		 * @parent can-connect/real-time/real-time.methods
 		 *
 		 * Programatically indicate a new instance has been updated.
 		 *
 		 * @signature `connection.updateInstance(props)`
 		 *
-		 *   Calls [can-connect/real-time.updatedData] in the right way so
+		 *   Calls [can-connect/real-time/real-time.updatedData] in the right way so
 		 *   that the instance is updated and added to or removed from
 		 *   any lists it belongs in.
 		 *
@@ -289,8 +289,8 @@ module.exports = connect.behavior("real-time",function(baseConnect){
 			});
 		},
 		/**
-		 * @function can-connect/real-time.destroyedData destroyedData
-		 * @parent can-connect/real-time.callbacks
+		 * @function can-connect/real-time/real-time.destroyedData destroyedData
+		 * @parent can-connect/real-time/real-time.callbacks
 		 *
 		 * @param {Object} props
 		 * @param {Object} params
@@ -307,15 +307,15 @@ module.exports = connect.behavior("real-time",function(baseConnect){
 			return undefined;
 		},
 		/**
-		 * @function can-connect/real-time.destroyInstance destroyInstance
-		 * @parent can-connect/real-time.methods
+		 * @function can-connect/real-time/real-time.destroyInstance destroyInstance
+		 * @parent can-connect/real-time/real-time.methods
 		 *
 		 * Programatically indicate a new instance has been destroyed.
 		 *
 		 * @signature `connection.destroyInstance(props)`
 		 *
 		 *   Gets or creates an instance from `props` and uses
-		 *   it to call [can-connect/real-time.destroyedData]
+		 *   it to call [can-connect/real-time/real-time.destroyedData]
 		 *   correctly.
 		 *
 		 * @param {Object} props The properties of the destroyed instance.
